Add file rename, use formatter

+++ src/LuaOnArduino.cpp
@@ -1,7 +1,7 @@
 #include <LuaOnArduino.h>
 
-#include <LuaLogLibrary.h>
 #include <LuaHmrLibrary.h>
+#include <LuaLogLibrary.h>
 
 LuaOnArduino::LuaOnArduino(SLIPSerial *slipSerial) {
   this->slipSerial = slipSerial;
@@ -20,13 +20,9 @@ void LuaOnArduino::begin() {
     if (that->oscInputHandler != NULL) that->oscInputHandler(oscInput);
   });
 
-  bridge.onRawInput([](uint8_t c) {
-    that->fileTransfer.write(c);
-  });
+  bridge.onRawInput([](uint8_t c) { that->fileTransfer.write(c); });
 
-  bridge.onRawInputEnd([]() {
-    that->fileTransfer.endWriteFile();
-  });
+  bridge.onRawInputEnd([]() { that->fileTransfer.endWriteFile(); });
 
   lua.onErrorBegin([]() { that->logger.errorBegin(); });
   lua.onErrorEnd([]() { that->logger.logEnd(); });
@@ -53,9 +49,7 @@ void LuaOnArduino::onInstall(CustomInstaller installer) {
   customInstaller = installer;
 }
 
-void LuaOnArduino::beforeReset(ResetHandler handler) {
-  handleReset = handler;
-}
+void LuaOnArduino::beforeReset(ResetHandler handler) { handleReset = handler; }
 
 void LuaOnArduino::onOscInput(OscInputHandler handler) {
   oscInputHandler = handler;
@@ -63,17 +57,23 @@ void LuaOnArduino::onOscInput(OscInputHandler handler) {
 
 void LuaOnArduino::update() { bridge.update(); }
 
-
 /**
  * Dispatch incoming OSC messages.
  */
 void LuaOnArduino::handleOscInput(OSCMessage &oscInput) {
-  static LuaOnArduino* that = this;
+  static LuaOnArduino *that = this;
   static char dirName[maxFileNameLength];
   static char fileName[maxFileNameLength];
 
   oscInput.dispatch("/read-file", [](OSCMessage &message) {
     uint16_t responseId = message.getInt(0);
+
+    if (!message.isString(1)) {
+      that->bridge.sendResponse(
+          Bridge::ResponseError, responseId, "Wrong argument");
+      return;
+    }
+
     message.getString(1, fileName, maxFileNameLength);
     that->fileTransfer.readFile(fileName, responseId);
   });
@@ -81,8 +81,13 @@ void LuaOnArduino::handleOscInput(OSCMessage &oscInput) {
   oscInput.dispatch("/write-file", [](OSCMessage &message) {
     uint16_t responseId = message.getInt(0);
 
+    if (!message.isString(1) || !message.isString(2)) {
+      that->bridge.sendResponse(
+          Bridge::ResponseError, responseId, "Wrong arguments");
+      return;
+    }
+
     message.getString(1, dirName, maxFileNameLength);
-    // This is the file name without directory.
     message.getString(2, fileName, maxFileNameLength);
 
     // Start reading the raw serial input. The readSerialMode is switched back
@@ -91,34 +96,69 @@ void LuaOnArduino::handleOscInput(OSCMessage &oscInput) {
     that->fileTransfer.startWriteFile(dirName, fileName, responseId);
   });
 
-  oscInput.dispatch("/delete-file", [](OSCMessage &message){
+  oscInput.dispatch("/delete-file", [](OSCMessage &message) {
     uint16_t responseId = message.getInt(0);
+
+    if (!message.isString(1)) {
+      that->bridge.sendResponse(
+          Bridge::ResponseError, responseId, "Wrong argument");
+      return;
+    }
+
     message.getString(1, fileName, maxFileNameLength);
     that->fileTransfer.deleteFile(fileName, responseId);
   });
 
   oscInput.dispatch("/create-dir", [](OSCMessage &message) {
     uint16_t responseId = message.getInt(0);
+
+    if (!message.isString(1)) {
+      that->bridge.sendResponse(
+          Bridge::ResponseError, responseId, "Wrong argument");
+      return;
+    }
+
     message.getString(1, dirName, maxFileNameLength);
     that->fileTransfer.createDirectory(dirName, responseId);
   });
 
   oscInput.dispatch("/delete-dir", [](OSCMessage &message) {
     uint16_t responseId = message.getInt(0);
+
+    if (!message.isString(1)) {
+      that->bridge.sendResponse(
+          Bridge::ResponseError, responseId, "Wrong argument");
+      return;
+    }
+
     message.getString(1, dirName, maxFileNameLength);
     that->fileTransfer.deleteDirectory(dirName, responseId);
-  });     
+  });
 
   oscInput.dispatch("/list-dir", [](OSCMessage &message) {
     uint16_t responseId = message.getInt(0);
+
+    if (!message.isString(1)) {
+      that->bridge.sendResponse(
+          Bridge::ResponseError, responseId, "Wrong argument");
+      return;
+    }
+
     message.getString(1, dirName, maxFileNameLength);
     that->fileTransfer.listDirectory(dirName, responseId);
   });
 
   oscInput.dispatch("/lua/run-file", [](OSCMessage &message) {
     uint16_t responseId = message.getInt(0);
+
+    if (!message.isString(1)) {
+      that->bridge.sendResponse(
+          Bridge::ResponseError, responseId, "Wrong argument");
+      return;
+    }
+
     message.getString(1, fileName, maxFileNameLength);
-    
+
     if (that->lua.runFile(fileName)) {
       that->bridge.sendResponse(Bridge::ResponseSuccess, responseId);
     } else {
@@ -128,11 +168,19 @@ void LuaOnArduino::handleOscInput(OSCMessage &oscInput) {
 
   oscInput.dispatch("/lua/update-file", [](OSCMessage &message) {
     uint16_t responseId = message.getInt(0);
+
+    if (!message.isString(1)) {
+      that->bridge.sendResponse(
+          Bridge::ResponseError, responseId, "Wrong argument");
+      return;
+    }
+
     message.getString(1, fileName, maxFileNameLength);
 
     bool usedHmr = LuaHmrLibrary::updateFile(fileName);
     if (!usedHmr) that->reset(true);
 
-    that->bridge.sendResponse(Bridge::ResponseSuccess, responseId, (int)usedHmr);
+    that->bridge.sendResponse(
+        Bridge::ResponseSuccess, responseId, (int)usedHmr);
   });
 }
